+++ cmd/empire/factories.go
@@ -7,6 +7,8 @@ import (
 	"net/url"
 	"os"
 
+	"golang.org/x/net/context"
+
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/client"
 	"github.com/aws/aws-sdk-go/aws/session"
@@ -17,6 +19,7 @@ import (
 	"github.com/remind101/empire"
 	"github.com/remind101/empire/events/app"
 	"github.com/remind101/empire/events/sns"
+	"github.com/remind101/empire/events/stdout"
 	"github.com/remind101/empire/pkg/dockerauth"
 	"github.com/remind101/empire/pkg/dockerutil"
 	"github.com/remind101/empire/pkg/ecsutil"
@@ -24,9 +27,30 @@ import (
 	"github.com/remind101/empire/scheduler"
 	"github.com/remind101/empire/scheduler/cloudformation"
 	"github.com/remind101/empire/scheduler/ecs"
+	"github.com/remind101/pkg/logger"
 	"github.com/remind101/pkg/reporter"
 	"github.com/remind101/pkg/reporter/hb"
 )
+
+// newRootContext returns a new root context.Context with an error reporter and
+// logger embedded in the context.
+func newRootContext(c *cli.Context) (context.Context, error) {
+	r, err := newReporter(c)
+	if err != nil {
+		return nil, err
+	}
+	l := newLogger()
+
+	ctx := context.Background()
+	if r != nil {
+		ctx = reporter.WithReporter(ctx, r)
+	}
+	if l != nil {
+		ctx = logger.WithLogger(ctx, l)
+	}
+
+	return ctx, nil
+}
 
 // DB ===================================
 
@@ -38,11 +62,6 @@ func newDB(c *cli.Context) (*empire.DB, error) {
 
 func newEmpire(db *empire.DB, c *cli.Context) (*empire.Empire, error) {
 	docker, err := newDockerClient(c)
-	if err != nil {
-		return nil, err
-	}
-
-	reporter, err := newReporter(c)
 	if err != nil {
 		return nil, err
 	}
@@ -70,14 +89,12 @@ func newEmpire(db *empire.DB, c *cli.Context) (*empire.Empire, error) {
 	e := empire.New(db, empire.Options{
 		Secret: c.String(FlagSecret),
 	})
-	e.Reporter = reporter
 	e.Scheduler = scheduler
 	if logs != nil {
 		e.LogsStreamer = logs
 	}
 	e.EventStream = empire.AsyncEvents(streams)
 	e.ProcfileExtractor = empire.PullAndExtract(docker)
-	e.Logger = newLogger()
 	e.Environment = c.String(FlagEnvironment)
 	e.RunRecorder = runRecorder
 	e.MessagesRequired = c.Bool(FlagMessagesRequired)
@@ -283,6 +300,12 @@ func newEventStreams(c *cli.Context) (empire.MultiEventStream, error) {
 			return streams, err
 		}
 		streams = append(streams, e)
+	case "stdout":
+		e, err := newStdoutEventStream(c)
+		if err != nil {
+			return streams, err
+		}
+		streams = append(streams, e)
 	default:
 		e := empire.NullEventStream
 		streams = append(streams, e)
@@ -311,6 +334,12 @@ func newSNSEventStream(c *cli.Context) (empire.EventStream, error) {
 	log.Println("Using SNS events backend with the following configuration:")
 	log.Println(fmt.Sprintf("  TopicARN: %s", e.TopicARN))
 
+	return e, nil
+}
+
+func newStdoutEventStream(c *cli.Context) (empire.EventStream, error) {
+	e := stdout.NewEventStream(newConfigProvider(c))
+	log.Println("Using Stdout events backend")
 	return e, nil
 }
 
@@ -348,7 +377,7 @@ func newLogger() log15.Logger {
 func newReporter(c *cli.Context) (reporter.Reporter, error) {
 	u := c.String(FlagReporter)
 	if u == "" {
-		return empire.DefaultReporter, nil
+		return nil, nil
 	}
 
 	uri, err := url.Parse(u)
@@ -372,7 +401,7 @@ func newHBReporter(key, env string) (reporter.Reporter, error) {
 
 	// Append here because `go vet` will complain about unkeyed fields,
 	// since it thinks MultiReporter is a struct literal.
-	return append(reporter.MultiReporter{}, empire.DefaultReporter, r), nil
+	return append(reporter.MultiReporter{}, reporter.NewLogReporter(), r), nil
 }
 
 // Auth provider =======================